requirements.txt
wikipedia-api
google-genai
IPython